Use the STL hash function for Key

+++ src/directory.cpp
@@ -30,11 +30,11 @@ Data Directory::Get( const Key& key ) const
 //
 std::size_t Directory::GetEntryId( const Key& key ) const
 {
-    std::uint32_t mask = 0;
+    std::size_t mask = 0;
     for( std::uint32_t i = 0; i < m_globalDepth; i++ )
         mask |= ( 1 << i );
 
-    std::uint32_t hv = key.GetHash();
+    std::size_t hv = key.GetHash();
 
     std::size_t ret = hv & mask;
     assert( ret < m_dir.size() );
